Created ProcessId (#97)
* refactor: created ProcessId and moved process to its own module

* style: ran cargo fmt

* refactor: now using ProcessId everywhere instead of a simple string and moved log related functions in its own module

* revert/fix: now printing back the whole process identifier (name + id) just like before

* refactor: deleted unnessary allocation

* refactor: moved process out of process directory

+++ taskmaster/src/tasks_manager/routine/handle_command/reload.rs
@@ -217,7 +217,7 @@ mod tests {
 
         let (mut routine, _status_receiver, _log_receiver) = test_routine(&current_config).await;
 
-        let unchanged_ids_before: Vec<u64> = routine
+        let unchanged_ids_before: Vec<usize> = routine
             .processes
             .as_inner()
             .lock()
@@ -227,7 +227,7 @@ mod tests {
             .iter()
             .map(|process| process.instance_id())
             .collect();
-        let changed_ids_before: Vec<u64> = routine
+        let changed_ids_before: Vec<usize> = routine
             .processes
             .as_inner()
             .lock()
@@ -353,7 +353,7 @@ mod tests {
         let (mut routine, _status_receiver, _log_receiver) = test_routine(&current_config).await;
 
         // initial state: 2 procs
-        let before_ids: Vec<u64> = routine
+        let before_ids: Vec<usize> = routine
             .processes
             .as_inner()
             .lock()
@@ -413,7 +413,7 @@ mod tests {
         let (mut routine, _status_receiver, _log_receiver) = test_routine(&current_config).await;
 
         // initial state: 2 procs
-        let before_ids: Vec<u64> = routine
+        let before_ids: Vec<usize> = routine
             .processes
             .as_inner()
             .lock()
